correcting phase 1 and 2

+++ src/WAVTrack.cpp
@@ -14,7 +14,7 @@ void WAVTrack::load() {
     // TODO: Implement realistic WAV loading simulation
     // NOTE: Use exactly 2 spaces before the arrow (→) character
     std::cout << "[WAVTrack::load] Loading WAV: \"" << title << "\" at " <<sample_rate<< "Hz/" <<bit_depth<< "bit (uncompressed)...\n";
-    long long size = static_cast<long long> (duration_seconds) * (sample_rate) * (bit_depth / 8) * 2;
+    long long size = duration_seconds * sample_rate * (bit_depth / 8) * 2;
     std::cout << "  → Estimated file size:" << size << "bytes \n";
     std::cout << "  → Fast loading due to uncompressed format. \n";
 }
@@ -28,11 +28,9 @@ void WAVTrack::analyze_beatgrid() {
     // 3. Print number of beats and mention uncompressed precision
     // should print "  → Estimated beats: <beats>  → Precision factor: 1.0 (uncompressed audio)"
 
-    double beats_estimated = static_cast<double> (((duration_seconds) / 60.0) * bpm);
-
-    int precision_factor = 1;
-
-    std::cout << "  → Estimated beats: " << beats_estimated << "  → Precision factor: " << precision_factor << " (uncompressed audio) \n";
+    double beats_estimated = ((duration_seconds) / 60.0) * bpm;
+    double precision_factor = 1;
+    std::cout << "  → Estimated beats: " << (int)beats_estimated << "  → Precision factor: " << precision_factor << " (uncompressed audio) \n";
 
 }
 
@@ -43,15 +41,15 @@ double WAVTrack::get_quality_score() const {
 
     double base_score = 70;
 
-    if(sample_rate >= 44100){
+    if (sample_rate >= 44100){
         base_score = base_score + 10;
     }
 
-    if(sample_rate >= 96000){
+    if (sample_rate >= 96000){
         base_score = base_score + 5;
     }
 
-    if(bit_depth >= 16){
+    if  (bit_depth >= 16){
         base_score = base_score + 10;
     }
 
@@ -59,8 +57,7 @@ double WAVTrack::get_quality_score() const {
         base_score = base_score + 5;
     }
 
-
-    if(base_score > 100){
+    if (base_score > 100){
         base_score = 100;
     }
     return base_score;
